Move features extraction to structs package
